04.04.2022 Delivery Добавлено предупреждение при удалении курьера
