+++ src/components/CourierStatusChange/CourierStatusChange.js
@@ -5,10 +5,10 @@ import {useState} from "react";
 import {useDispatch, useSelector} from "react-redux";
 import {changeCourier} from "../../store/couriers/actions";
 import {selectRequestCouriers} from "../../store/couriers/selector";
-//import {selectRequestCouriers, selectUpdateStatus} from "../../store/couriers/selector;
 
 export const CourierStatusChange = (props) => {
     const courier = props?.courier[0];
+    const courierID = +localStorage.getItem('id_user');
     const statusId = courier?.user_status_id - 1;
     const [checked, setChecked] = useState(!!statusId);
     const dispatch = useDispatch();
@@ -17,15 +17,17 @@ export const CourierStatusChange = (props) => {
     if (statusId === 1) label = 'Online';
     else if (statusId === 2) label = 'Busy';
     else label = 'Offline';
+    console.log('CourierStatusChange', statusId, checked, label);
 
     const requestCouriers = useSelector(selectRequestCouriers);
 
-    const onChangeStatus = (courier_id, event) => {
-        console.log('onChangeStatus', event)
-        dispatch(changeCourier({id: courier_id, user_status_id: event}));
+    const onChangeStatus = (courier_id, user_status_id) => {
+        console.log('onChangeStatus',courier_id,  user_status_id)
+        dispatch(changeCourier({id: courier?.id, user_status_id: user_status_id}));
     };
 
     const handleChange = (event) => {
+        console.log('handleChange', event.target.checked + 1)
         onChangeStatus(courier?.id, event.target.checked + 1);
         if (requestCouriers.error === null) setChecked(event.target.checked);
     };
