+++ src/components/CouriersPage/CouriersPage.js
@@ -2,16 +2,12 @@ import * as React from "react";
 import {useEffect, useState} from "react";
 import Typography from '@mui/material/Typography';
 import Stack from '@mui/material/Stack';
-import {Grid} from "@mui/material";
-import {Route, Routes, useParams} from "react-router-dom";
-import MyMap from "../Map/map";
-import CourierOrder from "../CourierOrder/CourierOrder";
+import {Route, Routes} from "react-router-dom";
 import {CourierStatusChange} from "../CourierStatusChange/CourierStatusChange";
 import {Box} from "@mui/material";
 import Menu from "../../utils/Menu";
 import {CourierMenu} from "../CourierMenu/CourierMenu";
 import {useDispatch, useSelector} from "react-redux";
-import {selectCurrentCourier} from "../../store/couriers/selector";
 import {
     selectDeliveredOrdersForCourier,
     selectTransitOrderForCourier
@@ -20,22 +16,19 @@ import CourierHistory from "../CourierHistory/CourierHistory";
 import {Chat} from "../Chat/Chat";
 import {getOrders} from "../../store/orders/actions";
 import {getCouriers} from "../../store/couriers/actions";
-import {AdminMain} from "../AdminTable/AdminMain/AdminMain";
-import {CouriersOperation} from "../AdminTable/CouriersOperation/CouriersOperation";
-import {CourierRegistration} from "../AdminTable/CourierRegistration/CourierRegistration";
 import CourierMain from "../CourierMain/CourierMain";
+import {selectCurrentCourier} from "../../store/couriers/selector";
 
 
 const CouriersPage = () => {
-    //const courierID = +useParams().id;
     const courierID = +localStorage.getItem('id_user');
     const currentCourier = useSelector((state) => selectCurrentCourier(state, courierID) );
     const currentOrder = useSelector((state) => selectTransitOrderForCourier(state, courierID));
     const deliveredOrders = useSelector((state) => selectDeliveredOrdersForCourier(state, courierID));
-    console.log('courier', courierID, currentCourier, currentOrder, deliveredOrders)
-
     const courierBusy = currentOrder?.length > 0;
+    console.log('courierBusy', courierBusy, currentOrder?.length > 0)
 
+    //console.log('courier', courierBusy, courierID, currentCourier, currentOrder, deliveredOrders)
     const dispatch = useDispatch();
     useEffect((event) => {
         console.log('useEffect')
@@ -75,46 +68,6 @@ const CouriersPage = () => {
                 <Route path="Chat" element={<Chat  mode="Courier" currentCourier={currentCourier[0]} />} />
                 <Route path="CourierHistory" element={<CourierHistory orders={deliveredOrders}/>} />
             </Routes>
-
-            {/*{(option === '1')*/}
-            {/*    ? <Chat mode="Courier" currentCourier={currentCourier[0]}/>*/}
-            {/*    :(option === '0')*/}
-            {/*         ? (clickOnMapToggle*/}
-            {/*            ? <Box xs={{width: '100%'}}>*/}
-            {/*                <MyMap name={''} orders={[currentOrder[0]]} couriers={currentCourier}  clickOnMap={clickOnMap}/>*/}
-            {/*             </Box>*/}
-            {/*            :*/}
-            {/*             <>*/}
-            {/*                 <Stack sx={{mb: 5}} direction="column" spacing={2}>*/}
-            {/*                     <Typography sx={{mt: 2, mb: 4}} variant="h4" component="div" >*/}
-            {/*                        {currentCourier[0]?.name}*/}
-            {/*                         <span className="courier-status"></span>*/}
-            {/*                     </Typography>*/}
-            {/*                     <CourierStatusChange  courier={currentCourier} courierBusy={courierBusy}/>*/}
-            {/*                 </Stack>*/}
-
-            {/*                  <Stack direction="row" spacing={2}>*/}
-            {/*                     <Grid container>*/}
-            {/*                         <CourierOrder order={currentOrder[0]}/>*/}
-            {/*                         <Grid item xs={6}>*/}
-            {/*                             <Box sx={{}}>*/}
-            {/*                                 <MyMap name={''}*/}
-            {/*                                        orders={currentOrder}*/}
-            {/*                                        couriers={currentCourier}*/}
-            {/*                                        clickOnMap={clickOnMap}*/}
-            {/*                                        sizeWidth={'100%'}*/}
-            {/*                                        sizeHeight={'250px'} />*/}
-            {/*                             </Box>*/}
-            {/*                         </Grid>*/}
-            {/*                     </Grid>*/}
-            {/*                  </Stack>*/}
-
-            {/*                  <CourierHistory orders={deliveredOrders} onClick={() => setOption('2')}/>*/}
-            {/*             </>)*/}
-            {/*        :(option === '2')*/}
-            {/*            ? <CourierHistory orders={deliveredOrders}/>*/}
-            {/*            : <> </>*/}
-            {/* }*/}
         </>
     );
 };
