+++ src/components/Messages/Message/Message.js
@@ -3,9 +3,6 @@ import React from 'react';
 
 import {Avatar, Divider, Grid, List, ListItem, ListItemIcon, TextField, ListItemText } from "@mui/material";
 
-
-   
-
 const Message = ({text, author, id, avatar}) => {
     console.log(text, author, id)
 
@@ -32,7 +29,6 @@ const Message = ({text, author, id, avatar}) => {
             <Divider />
             </>
         )
-
-}
+};
 
 export default Message;
